feat: add phase-1 static export build split and blockers report
